Neighborhood Pocket Version 1.2
In this commit, Google map function has been implemented. The user can
click on the Report Button and a marker will be at the user current
location. A List View has been added underneath the map screen; this
List View maybe used like to display our news feed. Also, a Map will be
displayed if the user taps on the Map Button.