Neighborhood Pocket V 1.2.4
The dialog classes were all made inner classes of the ReportMap
Activity, that way communication was easier and the dialogs can call
functions of the activity. Also, in the last dialog where the user can
take a picture, the button that allows this now works. Currently, the
picture taken by the application returns the image as bitmap, but this
can be changed once the database is implemented.