Neighborhood Pocket V 2.0
The take picture feature is working completely. Also you can enter a
description and it will save. You can click on an item from list view
and see a detailed description of the suspicious activity. The picture
you take will be saved to the suspicious activity. Please test. The only
thing left is persistence from the database, and some aesthetic
tweaking.